+++ src/sentry/seer/endpoints/seer_rpc.py
@@ -90,7 +90,7 @@ from sentry.seer.fetch_issues.fetch_issues_given_exception_type import (
     get_latest_issue_event,
 )
 from sentry.seer.seer_setup import get_seer_org_acknowledgement
-from sentry.sentry_apps.webhooks import broadcast_webhooks_for_organization
+from sentry.sentry_apps.tasks.sentry_apps import broadcast_webhooks_for_organization
 from sentry.silo.base import SiloMode
 from sentry.snuba.referrer import Referrer
 from sentry.utils import snuba_rpc
@@ -642,17 +642,14 @@ def send_seer_webhook(*, event_name: str, organization_id: int, payload: dict) -
     if not features.has("organizations:seer-webhooks", organization):
         return {"success": False, "error": "Seer webhooks are not enabled for this organization"}
 
-    try:
-        broadcast_webhooks_for_organization(
-            resource_name="seer",
-            event_name=event_name,
-            organization_id=organization_id,
-            payload=payload,
-        )
-        return {"success": True}
-    except Exception as e:
-        logger.exception("Failed to send seer webhook: %s", e)
-        return {"success": False, "error": str(e)}
+    broadcast_webhooks_for_organization.delay(
+        resource_name="seer",
+        event_name=event_name,
+        organization_id=organization_id,
+        payload=payload,
+    )
+
+    return {"success": True}
 
 
 seer_method_registry: dict[str, Callable[..., dict[str, Any]]] = {
